add some error verification

+++ app/controller/coreController.js
@@ -1,3 +1,6 @@
+const errorHandler = require('../../service/errorService/errorHandler')
+
+
 /**
  * @module coreController
  * @description Le module coreController contient des fonctions qui effectuent des opérations CRUD de base sur une table de base de données donnée.
@@ -16,14 +19,25 @@ const coreController = {
        * @async
        * @param {Object} req - L'objet de requête entrant.
        * @param {Object} res - L'objet de réponse sortant.
+       * @param {Object} next - Pour permettre de passer au middleware suivant
        * @description Récupère tous les enregistrements de la table spécifiée et retourne le résultat sous forme d'objet JSON.
        */
-      getAll: async (req, res) => {
+      getAll: async (req, res, next) => {
         try {
+
           const result = await model.getAll(table);
-          res.json(result);
+
+          if (result) {
+
+            res.json(result);
+
+          }else {
+
+            errorHandler._204(req, res, next);
+          }
         } catch (e) {
-          console.log(e.error);
+          
+          errorHandler._500(req, res, next);
         }
       },
 
@@ -32,15 +46,30 @@ const coreController = {
        * @async
        * @param {Object} req - L'objet de requête entrant.
        * @param {Object} res - L'objet de réponse sortant.
+       * @param {Object} next - Pour permettre de passer au middleware suivant
        * @description Récupère un seul enregistrement de la table spécifiée en fonction de l'identifiant dans les paramètres de la requête et retourne le résultat sous forme d'objet JSON.
        */
-      getById: async (req, res) => {
+      getById: async (req, res, next) => {
         try {
-          const id = Number(req.params.id);
-          const result = await model.getById(table, id);
-          res.json(result);
+
+          if (req.params.id) {
+
+            const id = Number(req.params.id);
+            const result = await model.getById(table, id);
+
+            if(result) {
+
+              res.json(result);
+            }else {
+              errorHandler._204(req, res, next)
+            }
+
+          } else {
+            errorHandler._400(req, res, next)
+          }
+
         } catch (e) {
-          console.log(e.error);
+          errorHandler._500(req, res, next);
         }
       },
 
@@ -49,16 +78,31 @@ const coreController = {
        * @async
        * @param {Object} req - L'objet de requête entrant.
        * @param {Object} res - L'objet de réponse sortant.
+       * @param {Object} next - Permet de passer au middleware suivant
        * @description Supprime un seul enregistrement de la table spécifiée en fonction de l'identifiant dans les paramètres de la requête et retourne le résultat sous forme d'objet JSON.
        */
 
-      deleteById: async (req, res) => {
+      deleteById: async (req, res, next) => {
         try {
-          const id = req.params.id;
-          const result = await model.deleteById(table, id);
-          res.json(result);
+
+          if ( req.params.id) {
+
+            const id = req.params.id;
+            const result = await model.deleteById(table, id);
+
+            if(result) {
+
+              res.json(result);
+            } else {
+              errorHandler._204(req, res, next)
+            }
+
+          } else {
+            errorHandler._400(req, res, next)
+          }
+
         } catch (e) {
-          console.log(e.error);
+          errorHandler._500(req, res, next);
         }
       },
 
@@ -67,15 +111,24 @@ const coreController = {
        * @async
        * @param {Object} req - L'objet requête entrant.
        * @param {Object} res -  L'objet réponse sortant.
+       * @param {Object} next - Permet de passer au middleware suivant
        * @description Cette fonction permet de créer ou de mettre à jour un enregistrement unique dans la table spécifiée en fonction des données contenues dans le corps de la requête, puis renvoie le résultat sous forme d'objet JSON.
        */
-      createOrUpdate: async (req, res) => {
+      createOrUpdate: async (req, res, next) => {
         try {
+
+
           let data = req.body;
           const result = await model.createOrUpdate(table, data);
-          res.json(result);
+
+          if(result) {
+
+            res.json(result);
+          } else {
+            errorHandler._204(req, res, next);
+          }
         } catch (e) {
-          console.log("error", e);
+          errorHandler._500(req, res, next);
         }
       },
     };
